# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ClientServiceType =
  (typeof ClientServiceType)[keyof typeof ClientServiceType];

export const ClientServiceType = {
  bookkeeping: "bookkeeping",
  va: "va",
  hybrid: "hybrid",
} as const;

export interface Client {
  id: number;
  name: string;
  email: string;
  monthly_hour_budget: number;
  monthly_fee: number;
  service_type: ClientServiceType;
}

export type CreateClientInputServiceType =
  (typeof CreateClientInputServiceType)[keyof typeof CreateClientInputServiceType];

export const CreateClientInputServiceType = {
  bookkeeping: "bookkeeping",
  va: "va",
  hybrid: "hybrid",
} as const;

export interface CreateClientInput {
  name: string;
  email: string;
  monthly_hour_budget: number;
  monthly_fee: number;
  service_type: CreateClientInputServiceType;
}

export type ClientDashboardServiceType =
  (typeof ClientDashboardServiceType)[keyof typeof ClientDashboardServiceType];

export const ClientDashboardServiceType = {
  bookkeeping: "bookkeeping",
  va: "va",
  hybrid: "hybrid",
} as const;

export interface ClientDashboard {
  id: number;
  name: string;
  email: string;
  monthly_hour_budget: number;
  monthly_fee: number;
  service_type: ClientDashboardServiceType;
  hours_used_this_month: number;
  hours_remaining: number;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  pending: "pending",
  complete: "complete",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  client_id: number;
  assigned_to?: string | null;
  status: TaskStatus;
  due_date?: string | null;
  client_name?: string | null;
}

export interface CreateTaskInput {
  title: string;
  description?: string | null;
  client_id: number;
  assigned_to?: string | null;
  due_date?: string | null;
}

export type UpdateTaskInputStatus =
  (typeof UpdateTaskInputStatus)[keyof typeof UpdateTaskInputStatus];

export const UpdateTaskInputStatus = {
  pending: "pending",
  complete: "complete",
} as const;

export interface UpdateTaskInput {
  title?: string;
  description?: string | null;
  assigned_to?: string | null;
  status?: UpdateTaskInputStatus;
  due_date?: string | null;
}

export interface TimeEntry {
  id: number;
  client_id: number;
  task_id?: number | null;
  duration_minutes: number;
  date: string;
  client_name?: string | null;
  task_title?: string | null;
}

export interface CreateTimeEntryInput {
  client_id: number;
  task_id?: number | null;
  duration_minutes: number;
  date: string;
}

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  new: "new",
  contacted: "contacted",
  closed: "closed",
} as const;

export interface Lead {
  id: number;
  name: string;
  email?: string | null;
  estimated_value?: number | null;
  status: LeadStatus;
  lead_source?: string | null;
}

export type CreateLeadInputStatus =
  (typeof CreateLeadInputStatus)[keyof typeof CreateLeadInputStatus];

export const CreateLeadInputStatus = {
  new: "new",
  contacted: "contacted",
  closed: "closed",
} as const;

export interface CreateLeadInput {
  name: string;
  email?: string | null;
  estimated_value?: number | null;
  status?: CreateLeadInputStatus;
  lead_source?: string | null;
}

export type UpdateLeadInputStatus =
  (typeof UpdateLeadInputStatus)[keyof typeof UpdateLeadInputStatus];

export const UpdateLeadInputStatus = {
  new: "new",
  contacted: "contacted",
  closed: "closed",
} as const;

export interface UpdateLeadInput {
  name?: string;
  email?: string | null;
  estimated_value?: number | null;
  status?: UpdateLeadInputStatus;
  lead_source?: string | null;
}

export type ListTasksParams = {
  clientId?: number;
};

export type ListTimeEntriesParams = {
  clientId?: number;
};