# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ClientServiceType } from "./clientServiceType";

export interface Client {
  id: number;
  name: string;
  email: string;
  monthly_hour_budget: number;
  monthly_fee: number;
  service_type: ClientServiceType;
}