# lib/api-zod/src/generated/types/createLeadInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateLeadInputStatus } from "./createLeadInputStatus";

export interface CreateLeadInput {
  name: string;
  email?: string | null;
  estimated_value?: number | null;
  status?: CreateLeadInputStatus;
  lead_source?: string | null;
}