# lib/api-zod/src/generated/types/createTaskInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTaskInput {
  title: string;
  description?: string | null;
  client_id: number;
  assigned_to?: string | null;
  due_date?: string | null;
}