# lib/api-zod/src/generated/types/createTimeEntryInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTimeEntryInput {
  client_id: number;
  task_id?: number | null;
  duration_minutes: number;
  date: string;
}