# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from "./leadStatus";

export interface Lead {
  id: number;
  name: string;
  email?: string | null;
  estimated_value?: number | null;
  status: LeadStatus;
  lead_source?: string | null;
}