# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from "./taskStatus";

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  client_id: number;
  assigned_to?: string | null;
  status: TaskStatus;
  due_date?: string | null;
  client_name?: string | null;
}