# lib/db/src/schema/tasks.ts
import { pgTable, serial, text, integer } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const tasksTable = pgTable("tasks", {
  id: serial("id").primaryKey(),
  title: text("title").notNull(),
  description: text("description"),
  client_id: integer("client_id").notNull(),
  assigned_to: text("assigned_to"),
  status: text("status", { enum: ["pending", "complete"] }).notNull().default("pending"),
  due_date: text("due_date"),
});

export const insertTaskSchema = createInsertSchema(tasksTable).omit({ id: true });
export type InsertTask = z.infer<typeof insertTaskSchema>;
export type Task = typeof tasksTable.$inferSelect;